+++ source/extensions/health_checkers/grpc/BUILD
@@ -1,0 +1,31 @@
+load(
+    "//bazel:envoy_build_system.bzl",
+    "envoy_cc_extension",
+    "envoy_extension_package",
+)
+
+licenses(["notice"])  # Apache 2
+
+envoy_extension_package()
+
+envoy_cc_extension(
+    name = "health_checker_lib",
+    srcs = ["health_checker_impl.cc"],
+    hdrs = ["health_checker_impl.h"],
+    external_deps = [
+        "grpc_health_proto",
+    ],
+    extra_visibility = [
+        # previously considered core code.
+        "//test:__subpackages__",
+    ],
+    deps = [
+        "//source/common/grpc:codec_lib",
+        "//source/common/http:codec_client_lib",
+        "//source/common/upstream:health_checker_lib",
+        "//source/common/upstream:host_utility_lib",
+        "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
+        "@envoy_api//envoy/data/core/v3:pkg_cc_proto",
+        "@envoy_api//envoy/type/v3:pkg_cc_proto",
+    ],
+)
